agentic2-007/agentic2/app/requirements.txt
fastapi==0.109.1
openai>=1.0.0
uvicorn==0.27.0
pytest==8.3.5
httpx==0.26.0
python-dotenv==1.0.0